+++ src/components/ReFlop/index.ts
@@ -1,7 +1,7 @@
-import reFlop from "./src/index.vue";
+import _ReFlop from "./src/index.vue";
 import { withInstall } from "@pureadmin/utils";
 
 /** 时间翻牌组件 */
-export const ReFlop = withInstall(reFlop);
+export const ReFlop = withInstall(_ReFlop);
 
 export default ReFlop;
