+++ src/components/ReSplitPane/index.ts
@@ -1,0 +1,6 @@
+import _ReSplitPane from "./src";
+import { withInstall } from "@pureadmin/utils";
+
+export const ReSplitPane = withInstall(_ReSplitPane);
+
+export default ReSplitPane;
